CRIMAP-127 List of added offences
So we have something to "see" if the offences page is adding offences and dates as expected.

It is still kind of a mock (there is a presenter with some mock data), and no remove functionality implemented yet, that will be next in another PR, but the "change" link will work.