CRIMAP-130 Implement offence destroy action
Another follow-up to previous offence summary page PRs.

This one implements the "Remove" link.
It follows the same pattern we already did for the applications dashboard, with a confirmation page, and shows a flash banner as well.

I tweaked a bit the routes `crud_step` helper to simplify this and to expose the `confirm_destroy` action when we have a `destroy`.

As agreed, if the offence being removed is the last one in the list, we redirects to a brand new offence, but if there are more offences left, we redirect back to the summary page.