Add datalist for the non-JS offences dropdown
This datalist is populated from the CSV, and offences are assigned to the Charge as value-objects making use of `composed_of` where we find the most suitable offence by name and return it if found, otherwise we assume the offence is being "created" by the provider, in which case we will not have an `#offence` (returns nil) and thus we lack additional attributes like offence class and type.

Add basic presence validation on the `offence_name` attribute.

Update the `ChargePresenter` as almost everything is now retrieved from the offence instance (or returns `nil` for manual offences).