Simplify offence dates markup
By enclosing the `fields_for` (the loop so to speak) in a `fieldset`, we create a better hierarchy of form elements and it is easier to then have the individual date legends are all visually hidden, and only the first date hint os shown too.

Tidy up the locales.

Note: there is an inconsistent font size/boldness on offence name and offence dates legends, this will be sorted out separately as there is another open PR for offence name.