Present the offence class(es), comma separated
This is a tweak to adapt the way we show the offence class in the views, to the requirements of the design.

First, we unify the presentation of the class in just one single place (`OffencePresenter`). This presenter works with `Offence` instances thus always will have class attributes.

Additionally modify the existing `ChargePresenter` use the new `OffencePresenter` also to present information related to the offence, including the class.

Last, a new `OffenceCollectionDecorator` takes care of the offences dropdown, also presenting them with the new `OffencePresenter` so class is presented comma-separated as well.