Fix for undefined when clicking the offence class
There was a funny intermittent bug where sometimes selecting an offence from the suggestion resulted in an `undefined` offence name in the text input.

The culprit was caused by the introduction of the offence class. If you click the offence class we wouldn't have the attributes of the offence as those are in the `li` element.

This fixes the problem ensuring no matter where you click in the suggestions, we always obtain the `li` element.