CRIMAP-129 Enhancement of offence datalist with JS (#115)
* CRIMAP-129 Enhancement of offence datalist with JS

Preliminary discovery work and first rough version of the javascript enhancement in the offences `datalist`.

This produces a "suggestions" dropdown, with consistent look and feel across all browsers, when javascript is enabled.

When javascript is disabled the normal "datalist" is shown instead so functionality is not lost.

Fair disclaimers:

- We are piggybacking on a draft PR for an unfinished work, not part of the GOV.UK Frontend yet. And it might not be part of it for a long time, if ever.

- It just doesn't work out of the box. I had to made a few modifications to the javascript.

- Additionally, to show the offence class, I also had to extend the functionality with a `data-caption` feature. This has an impact on the render of the datalist markup too.

- I've marked all this modifications with `[change: xyz]` so it is easy to identify them.

Finally, for the sake of simplification and this initial version, I've removed the "Class Z, Y, Z" from the non-javascript datalist. Keeping it required additional changes to the javascript and I wanted to keep these modifications as light and limited as possible.

Some modifications I'm thinking contributing back to the draft PR.

* Fix for undefined when clicking the offence class

There was a funny intermittent bug where sometimes selecting an offence from the suggestion resulted in an `undefined` offence name in the text input.

The culprit was caused by the introduction of the offence class. If you click the offence class we wouldn't have the attributes of the offence as those are in the `li` element.

This fixes the problem ensuring no matter where you click in the suggestions, we always obtain the `li` element.