CRIMAP-75 Implement basic court hearing step
This includes the route, controller, view and form object as well as new required DB attributes on the `cases` table.

Court name input text, does not implement the dropdown or typeahead / suggestions yet. That will be part of a separate task/PR as it requires further research around where this data is coming from or how to "store" it in our application (if we do).

Co-Authored-By: timpeat <tim@timpeat.com>
Co-Authored-By: Chris Gannon <cjgannon@gmail.com>