Reset correspondence address when not applicable
As we do in other places, if provider changes the correspondence type selection from `other` to anything else (like home or provider's office) we reset the previously entered correspondence address (if any).

As this is a `has_one` relationship, nilling the attribute will remove the record from the `Addresses` table.