Add CODEOWNERS file
This uses a child team where I've added all of us developers.
After this is merged I will configure the repo settings to auto-request a PR review from this team.

For now I've added all of us (developers) to the child team but if any of us get to focus more on review instead of apply, we can edit the team members.