CRIMAP-163 Better handling of incomplete offences (#134)
Due to the way the offences page work (a page with nested attributes for details of the offence like offence name and one or more dates), each time the provider adds a new offence through the summary page, a blank record is created in the database, for the provider to fill in the details (offence name and dates).

This usually happens in one go, but if at that point, without finalising all the details, the provider clicks the “save and come back later“ button, for example, or click the “back“ link in the page, they will leave the offence in an incomplete state, maybe missing the name, or the dates, or both.

Removing these empty records is not a nice solution really, as some might contain partially entered details, like the offence name (but no dates) or vice versa, if the provider clicks the save draft button, in order to come back later on to these offences.

Also it breaks the navigation as clicking the "back" link in the summary page might go to a just deleted charge record.

A better solution is to present these incomplete charges for the provider to either remove them or edit and complete them.

Also some additional validation is added to not allow a provider to proceed with incomplete charges in the "basket" page.

Discussed the copy and design with service designer and we decided to go with the MOJ badges.

MOJ library is added to the package.json but only the badge component is used. In a separate PR I will also refactor the task list to use the MOJ import instead of maintaining our own copy of the component.