Import MOJ task-list component
As a follow-up to PR #134 where we introduced the MOJ design system in order to use the badges, now we can also replace the task-list styles with the MOJ component.

This is desirable so we get updates and fixes whenever they update the component and we pull a new version.

A few tweaks were needed however in order to render the tags properly but I've kept these as minimal and with separate class names to not clash with the MOJ ones.