CRIMAP-155 Review client details (#137)
Note: this PR is quite large as it sets out the "framework" to present the "check your answers" page and to edit the answers (change links).

As part of this framework, the `Client details` section has been partially implemented. There are a few remaining questions-answers I'm leaving for a follow-up PR so this one does not grow even larger.

Despite the many files, only 5 classes are the main implementation. The other files are view partials and many tests, including some view tests.

The general idea is: each time we want to add a new section to the check your answers page, we implement a new class inheriting from `Sections::BaseSection` and define all the answers making use of the different "answer" types and their change link, if any.

These classes will render themselves making use of partials. Each type of answer renders slightly different (for example a date answer will render the localised date, while a free text answer will render as it is).