CRIMAP-157 Review offences (#142)
Add the offences section to the check your answers page.

This renders a slightly different markup as the offences are indexed, and shows more details like the class or the dates.

In order to do this in a clean way, a specific "OffenceAnswer" component has been created, with its corresponding view partial. Other than that it behaves the same as other answers really.