CRIMAP-114 Add declaration page
Simple page, with content (may change, awaiting approval) and a check box to confirm, in order to proceed to what will be the final step, the submission confirmation (in a future PR).

Note: because this page is mostly all content, it makes sense to have it as a localised view (`edit.en.html.erb`) and not extract to locales the content as that would make it more difficult to update/translate in the future.

Linked this step to the overall journey, right after the review page.