Add IoJ task and logic
Also, compact the array of Ioj types to make sure we always store
a cleaned up array, with no empty strings that can result due to
the way the form builder submits collection of check boxes.

Simplify a bit the `validate_types` method, as we can safely assume
we always have an `array` (the attribute performs the casting),
and no need to do `compact_blank` anymore.