Refactor and finalise if possible the current tasks
Still some pending things, like the NINO check.

Reorganised the way the status is returned so we implicitly are protected against possible `nil` values (like case record not yet being created). This way we don't have to constantly in code use the safe navigation operator.

Additionally, no need to pass the crime_application record to the paths, it is retrieved from the `default_url_options`