Add Review and Declaration to the task list
These are the last 2 remaining tasks we have in the task list for MVP.

As the `Review` task does not really perform any update (it is just a page with change links) it will show as in progress as soon as the previous `Ioj` task is fulfilled. And it shows as completed if declaration has been fulfilled.

As for the `Declaration` task, it shows also as in progress as the `Review`, with same logic. But it shows as completed when the declaration has been signed.

As we don't have yet the logic for the submission, this is all for now. Once we implement the application "submission" and mark it as submitted, it becomes read only and there may be further changes needed to the task list (but this will be done as separate PRs).