CRIMAP-171 Add IoJ to task list and improve existing tasks logic (#152)
* Refactor and finalise if possible the current tasks

Still some pending things, like the NINO check.

Reorganised the way the status is returned so we implicitly are protected against possible `nil` values (like case record not yet being created). This way we don't have to constantly in code use the safe navigation operator.

Additionally, no need to pass the crime_application record to the paths, it is retrieved from the `default_url_options`

* Add IoJ task and logic

Also, compact the array of Ioj types to make sure we always store
a cleaned up array, with no empty strings that can result due to
the way the form builder submits collection of check boxes.

Simplify a bit the `validate_types` method, as we can safely assume
we always have an `array` (the attribute performs the casting),
and no need to do `compact_blank` anymore.