CRIMAP-113 Add submission confirmation page (#154)
This is just a show action, rendering a localised view (as there is no complex logic or anything like that, just some content).

Updated the decision tree, so this page is shown right after the declaration page.

In another PR, the green button "View completed application" will be implemented to show a read-only view of the "check your answers". At the moment, it just goes to the list of applications.

Note: for now there is no logic to mark applications as submitted (and of course, no actual "submission" to any place). This is all smoke and mirrors until we flesh out next pieces of the puzzle.