CRIMAP-170 View completed application
This is what in the prototype is called the application certificate and is a read-only view of the "review / check your answers" page.

The print button works but is not yet applying the print stylesheet, that will come as part of a separate ticket.

Also, for now I've had to make a lot of assumptions to get to this point, mainly, the "submission" which for not is no more than a mock service `ApplicationSubmission` that sets the `submitted` status in the record.

I've also added a new `submitted_at` DB field because it will be needed for some date stamp logic soon.

In separate PRs, we will introduce checks to make sure a "submitted" application can't be edited, or for what is worth, can't be re-submitted (which at the moment is possible as there is no such validation/check yet).