CRIMAP-176 Implement date stamp logic on submission
The date stamp logic is a bit more tricky than anticipated, but in summary, it follows these broad rules:

- If case type is “date stampable”, we set and use the date stamp value across the application
- If case type is non “date stampable”, we don't set it until submission, at which time this becomes the date stamp

Across the application there are places (mainly task list) where the date stamp needs to show even if the application is not yet submitted, so the logic has been improved to cope with this.

On the certificate page (post submission) we've added in addition to the date stamp, the submission date. For non-stampable case types these 2 dates will be identical. For stampable case types, the dates will differ.