Rename method to `interim_date_stamp`
I don't think the logic was clear enough and I now have make it more explicit by renaming the method #application_date_stamp to #interim_date_stamp and NOT using this method in the post-submission "certificate" page, and instead using whatever values are in the date_stamp and in the submitted_at as the source of truth.

What it means is, anything pre-submission (in progress applications) show an "interim" date stamp based on the logic whether a case type is date stampable or not.

And the `date_stamp` can be populated but due to a change in the case_type, then it is hidden from the UI (in the task list for example).

However, post submission, for "submitted" applications, this method is not used anymore, as we would have the definite values in both, the `date_stamp`, and the `submitted_at` attributes.