CRIMAP-176 Implement date stamp logic on submission (#159)
* CRIMAP-176 Implement date stamp logic on submission

The date stamp logic is a bit more tricky than anticipated, but in summary, it follows these broad rules:

- If case type is “date stampable”, we set and use the date stamp value across the application
- If case type is non “date stampable”, we don't set it until submission, at which time this becomes the date stamp

Across the application there are places (mainly task list) where the date stamp needs to show even if the application is not yet submitted, so the logic has been improved to cope with this.

On the certificate page (post submission) we've added in addition to the date stamp, the submission date. For non-stampable case types these 2 dates will be identical. For stampable case types, the dates will differ.

* Rename method to `interim_date_stamp`

I don't think the logic was clear enough and I now have make it more explicit by renaming the method #application_date_stamp to #interim_date_stamp and NOT using this method in the post-submission "certificate" page, and instead using whatever values are in the date_stamp and in the submitted_at as the source of truth.

What it means is, anything pre-submission (in progress applications) show an "interim" date stamp based on the logic whether a case type is date stampable or not.

And the `date_stamp` can be populated but due to a change in the case_type, then it is hidden from the UI (in the task list for example).

However, post submission, for "submitted" applications, this method is not used anymore, as we would have the definite values in both, the `date_stamp`, and the `submitted_at` attributes.