Clarify routes for dashboard listings
There is a separate ticket CRIMAP-173 to redesign the dashboard to have 3 tabs (subnavigation really).

As 2 of those tabs act over submitted applications, thus we want to centralise those on the `SubmittedApplicationsController` which will be the controller that "talks" to the document store (and will not use ActiveRecord), in the future.

We can accomplish this by using the `index` action, and providing an optional filtering parameter, `status` to filter by `submitted` (this should be the default) or by `returned` (which is a subset of the submitted applications).

I've not implemented the action or the view as that is part of the other ticket CRIMAP-173.