Renaming Submitted to Completed in controller and routes
Submitted was misleading as it refers to a `status` of the application. Other status being `returned`, and there are others.

However my intention was to refer to "Remote" applications, as in non-local anymore (because they will be in the document store).

We've decided to go with `Completed` to better capture these are completed applications, as in "the application is no longer in our database".

A completed application can have different statuses, like `returned`. These are filters to be applied, but the fact the application is completed remains.
When an application is put back into editable mode, it becomes a local application, in_progress, and is no longer completed until submission again.