CRIMAP-174 Refactor to enable post-submission (#161)
* CRIMAP-174 Refactor to enable post-submission

This is a bit of a refactor to more clearly separate the boundaries between "local" (and thus `in_progress`) applications, and "remote" (or submitted) applications, those that eventually will reside in the document store.

The first challenge is identifying where in the service we are using a "current_crime_application" record, because at the moment all come from the local DB.

This will not be the case in the future when we have the document store.

Some pages can't continue referring to this method expecting an application is in the local DB. For example the submission confirmation is a funny one because right on submission we will have marked as "submitted" the application and eventually purged it from database, so this page can't rely on a record on the DB.

Fortunately this page renders very little, just a reference, and the UUID remains on the route URL, so passing the reference as a query param, allows us to not rely on a record, and to also make this page bookmarkable as recommend by the design guidelines:
https://design-system.service.gov.uk/patterns/confirmation-pages/

> Some users will bookmark the confirmation page as a form of receipt. You should allow them to return to the page, whenever possible.

The other post-submission page we have currently is the "view certificate". This has been moved to its own controller where maybe more actions in the future will be added.
The `current_crime_application` then scopes the DB query to `submitted` only, but once we introduce the document store, all these ActiveRecord queries need to be transformed into an API client querying the document store.

Finally, if the application in our local DB is marked as `submitted` (in the future, once these are purged, they will not even exist in local DB) and the provider tries to access any route making use of the record (any journey page for example, or task list, etc.) they will be given an error, as per discussion in the ticket.

* Clarify routes for dashboard listings

There is a separate ticket CRIMAP-173 to redesign the dashboard to have 3 tabs (subnavigation really).

As 2 of those tabs act over submitted applications, thus we want to centralise those on the `SubmittedApplicationsController` which will be the controller that "talks" to the document store (and will not use ActiveRecord), in the future.

We can accomplish this by using the `index` action, and providing an optional filtering parameter, `status` to filter by `submitted` (this should be the default) or by `returned` (which is a subset of the submitted applications).

I've not implemented the action or the view as that is part of the other ticket CRIMAP-173.

* Renaming Submitted to Completed in controller and routes

Submitted was misleading as it refers to a `status` of the application. Other status being `returned`, and there are others.

However my intention was to refer to "Remote" applications, as in non-local anymore (because they will be in the document store).

We've decided to go with `Completed` to better capture these are completed applications, as in "the application is no longer in our database".

A completed application can have different statuses, like `returned`. These are filters to be applied, but the fact the application is completed remains.
When an application is put back into editable mode, it becomes a local application, in_progress, and is no longer completed until submission again.