CRIMAP-97 Add additional actions to developer tools (#163)
This is a first PR to add some additional actions to the developer tools (in the footer) that will be necessary to progress with the user story in CRIMAP-97 (amend application).

Mainly, I've added a "mark as returned" button that will only show when "viewing" a `submitted` application. This button will put it into a `returned` status which later on can be used to filter in the dashboard and to amend it through the certificate page. This will be done in a separate follow-up PR.

I've also added a quick action to fill up an application up to the DWP check (success) result page, as I kept forgetting myself which combinations of surname, dob and NINO pass the check (mock values).

Overall, reorganised a bit the developer tools into its own controller, leaving the `SessionsController` as might be used in the future when we have the login mechanism.