Configurable summary presenters, reuse in DWP journey (#164)
* Greater flexibility in summary presenters

First thing, decouple the views from having to use the `current_crime_application` to know if the change links should show or not. Instead, make it part of the configuration of each of the sections (`editable`), and propagate this to the views as `locals`.

Same if we want to hide the header of the section. Default is false, but we can pass `headless: true` to hide the header.

This makes for more flexible "section" components so we can reuse them in other places (like in the DWP confirm applicant details page).

Also tests are simpler and do not need to have the knowledge or concept of a "crime application" record.

* Make use of existing `ClientDetails` component

With the tweaks done in previous commit we can now reuse the `Summary::Sections::ClientDetails`, the same we use in the check your answers, for the DWP passporting journey too.

A bit of markup tidy up.