CRIMAP-97 Amend returned application
Follow-up to previous preparation PRs, this one implements a barebones "amend" action, with a button in the certificate page so the provider can put back an application into "editable" mode (`in_progress`) and make changes in order to submit again the application.

Note: there are so many implementation details we are still missing that is not worth it taking this further until we have perhaps the document store.

For now is a bit smoke and mirrors but it serve the purpose of getting an idea how the user experience will look like and we can test this with providers.

I'm awaiting confirmation regarding some of the time stamps on the application, if we have to keep the originals or reset on re-submission.
I think we might have to introduce a `parent_id` at some point, it will keep track of previously submitted applications to create kind of a history of all amendments, but didn't want to add that DB attribute until we have more clarification.