CRIMAP-98 Print application certificate (#167)
Apply a basic print stylesheet in the application certificate page, to remove superfluous visual elements like the header or footer, when printing the page.

The printing can be invoked through the "Print application" button (but this will only work with javascript enabled) or just by going to the browser menubar and selecting "File - Print".

We can tweak the print stylesheet if required, but for now it looks good enough. Also, in the future, this might be superseded by a PDF download (probably post-MVP as it is not a very high priority right now).