CRIMAP-204 Add USN sequence (serial) DB attribute
This will be later on used to replace our current (mocked) LAA reference.

This PR only adds the sequence and attribute, along with a looong comment explaining some gotchas.

Upon migrate, any existing applications in the DB will see their `usn` attribute populated based on the sequence, starting at `6_000_000`

But this attribute is not yet used anywhere. A separate PR will be raise to replace the LAA reference with this USN.