Retrieve submitted/returned applications from datastore
The dashboard is now also retrieving applications from the datastore, assuming the feature flag is turned on (default to false in `local` to not disrupt other development).

Note however: if the submitted/returned application not longer exists in the local database, trying to "open it" (to see the certificate page) will fail as this functionality is not yet developed.

At the moment, even with the feature flag turned on, we are not purging application for this very reason, but that will soon change.