CRIMAP-39 Add proper content to home page
In preparation for user test sessions, make the homepage more similar to what eventually will be (once we have post-login journey).

Note the view is localised, as is heavy on content, with no dynamic/complex components.

The `start button` will go to the dashboard, this is the same as going to the `developer tools` and clicking "your applications", but is better for user testing, as providers might get confused what this developer tools is about.

Once we have the Portal Login, the start button in reality will go to the selection of office account number. But we can't implement that yet.