A few minor tweaks
With a small change to the `#show?` method, to not chain the `ioj_passport` attribute, we simplify some testing.

The attribute will always be present as it is an attribute of the case itself, not a relationship like we had with IoJ before.

We can assume it will be of type array. If it were to be `nil` it means we've introduced a bug in our code and then is best to blow up and know about this so we can fix it.