Submission end to end with datastore (feature flag) (#175)
* Submission end to end with datastore (feature flag)

This is a first stab at an integration between Apply and Datastore for submitting applications. And end to end so to speak from a provider's point of view.
Note however there is no proper adapter to create the JSON document and I'm using a dirty, quick sampling of attributes.

It currently is behind a feature flag, disabled on local by default as to not disrupt any work or force developers to run a local datastore service (for now).

However is enabled on staging on purpose to validate some assumptions regarding connectivity, etc.

This uses a little http client gem made for the only purpose of wrapping the http API in a common interface that could be shared between Apply and Review eventually.

* Retrieve submitted/returned applications from datastore

The dashboard is now also retrieving applications from the datastore, assuming the feature flag is turned on (default to false in `local` to not disrupt other development).

Note however: if the submitted/returned application not longer exists in the local database, trying to "open it" (to see the certificate page) will fail as this functionality is not yet developed.

At the moment, even with the feature flag turned on, we are not purging application for this very reason, but that will soon change.

* Push USN to datastore

* Disable feature on staging for now

We have UT sessions coming soon and this is not ready to show.