Add granularity to `FeatureFlags` for test env (#182)
As we develop some code behind feature flags, like right now the datastore integration, we might not want yet to go ahead and write detailed tests for it.

One way to handle this was to use, for instance:

```ruby
allow(
        FeatureFlags
      ).to receive(:datastore_submission).and_return(double(enabled?: false))
```

But when you have many places where the feature flag might trigger, adding this to every single test (and then remember to remove it, etc.) can be tiresome.

Adding some granularity to the `HostEnv` and `FeatureFlags` mechanism to treat the `test` environment as any other environment so we can turn on/off features will make this hopefully a bit easier.

By default, local and test envs, if not explicitly declared on the feature flags configuration, will be turned on. To disable just declare the env.