Pagination proof of concept (dynamodb)
A limitation of DynamoDB is the pagination is not like a traditional pagination where you have pointers to each page through an offset and a limit.

Pagination with DynamoDB is really an "infinite scroll" pagination where you just get a "token" based on the last record returned in the previous page, so you can pass this token in the next request and continue loading records where you left.

This means you can only advance forwards, and not go to a specific page, just the immediate next page to the one you are.

There are ways to get back to the previous page if you track the previous token as query params, but that is all, only current page, previous and next are supported, the moment you go backwards more than 1 page, you don't have a previous token anymore.

This PoC is a limited version of a pagination, where user can advance through pages, and they can't go back, unless they reach the last page in which case a "< Start" link shows to go back to page 1.

It is done a bit quick and dirty to get the sense and idea. I'm not sure if pagination might be a deal breaker to _not_ use dynamodb (along with probably in the near future complex filters and sort columns)...