Another pagination proof of concept (dynamodb)
This is another kind of pagination, as a follow up to PR #184.

This one is a bit more complex on the markup side (probably a fair amount of DRY and tidy up can be done, but this is a quick PoC), but may be more useful to users.

This one allows to always move, from the page you are in, one page forward and one page backwards, however when going backwards only the immediate previous page can be accessed.

The benefit of this pagination is that it is "almost" like a traditional pagination, and has visual feedback of which page you are in, which is the next one, and which is the previous one. When there are no more previous pages (because of dynamo limitation) we show page 1.