Update code to work with latest schema
Now that we've started to validate the JSON payload against the schema on the datastore, some of the code was outdated.
Also the JSON responses have changed so we need to adapt some of the proof of concept code we had (like the paginators).

In addition, the presenters need to work with both, DB applications and JSON applications. I've started some work on this front but is far from perfect and definitely kind of ugly.