Use dry-struct for JSON applications
Eventually, Apply and Review will both be reading JSON applications, so it makes sense if we can come up with some common interface we can use in both places.

For us is a bit more convoluted as we also treat with DB (ActiveRecord) applications so we try to reuse presenters, but for Review it will almost all consist of "structs".

I've literally copy&paste some of these classes from Review and made minor adjustments to fit current (v0.1) schema. Also removed things we don't use on Apply.

This can be an exercise to evaluate how practical might be to share this in a gem for Apply and Review to use.

The gem might be a separate one, or perhaps it makes sense to be integrated in the http client, after all, both Apply and Review will be using the gem to "talk" to the datastore API and it might make sense for the resulting objects of these calls to not be just plan JSON objects but these "dry-structs".

Up to discussion and more thinking...