CRIMAP-213 Render submitted application certificate (#196)
Ensure the already existing presenters we had (all under `app/presenters/summary/...`)  works also with a submitted application, as in an application coming from the Datastore and thus not persisted to Apply local DB.

It mostly works out of the box but a few tweaks were needed, because some attributes are named differently in our DB and in the datastore JSON.

To accomplish this in the most transparent way possible a few Adapters have been created in `app/services/adapters/structs/...`. The purpose of these adapters is to take the JSON from the datastore, create Struct objects (from the `laa_crime_schemas` gem) and then apply a thin layer on top of them to customise behaviour and allow for some method renames.

This is not yet completed, but in order to not make this PR too big, will leave those for another separate PR.
However this is perfectly functional at the moment (assuming in the datastore there are applications conforming to the latest version of the schema).