CRIMAP-213 Show offences in submitted certificate
Follow-up to previous PR #196 to finalise the `offences` section.

Because the way the presenters work internally to present the offences class, etc. in the different parts of the service (offences basket, check your answers, certificate page) there is a bit more of complexity in the "thin adapter layer", as we want to reuse our "Charge" model and not to re-implement most of it.

Also, charges in our database might not be "completed" (lacking some details like the name or the dates) but those coming from the datastore can be assumed to be completed as they fulfilled the JSON schema already.