CRIMAP-213 Refactor and show IoJ passported in certificate page
This is I believe the final bit of information we have to read from the datastore application and be able to render in the certificate page.

I've moved the DB attribute from the `case` to the `crime_application` table as it is something that applies globally to the "application" as a whole.

It also makes the code much simpler and I've been able to cleanup quite a lot and refactor some additional code in the task list page, and check your answers page.