Enable datastore in tests (webmock stub)
We are reaching to a point where we have quite a lot of functionality that we can start implementing tests to rely on having a "datastore" with endpoints to call.

I've adapted any scenarios in `spec/requests/dashboard_spec.rb` that should talk to the datastore, to stub the requests and responses so it no longer goes to the database and we don't have to create `CrimeApplication` records locally.

Slowly removing some conditional code, but more to do in next PRs.

I've also removed v1 of the paginator PoC. For now I've left v2 as some of it might be salvaged once we move to postgres on the datastore.